windows stuff

+++ device/device_control.hpp
@@ -208,7 +208,8 @@ namespace device {
 
         detail::cast_to_data_header(data)->type = proto::data_type_t::SCRIPT_NAME;
 
-        strncpy(reinterpret_cast<char*>(detail::cast_to_data(data)), name.data(), std::min(name.size(), 32UL));
+        strncpy(reinterpret_cast<char*>(detail::cast_to_data(data)), name.data()
+            , std::min(static_cast<size_t>(name.size()), (size_t)32U));
 
         send(data);
     }
